+++ src/app/(core)/teachers/programmes/[slug]/units/[unitSlug]/lessons/[lessonSlug]/Components/getLessonResources.tsx
@@ -103,7 +103,6 @@ const getSkipLinkUrl = ({
 export function getLessonResources({
   browsePathwayData,
   data,
-  copyrightState,
   isMathJaxLesson,
   trackMediaClipsButtonClicked,
   contentRestricted,
@@ -167,18 +166,6 @@ export function getLessonResources({
       displayVocab={!!data.actions?.displayVocabButton}
       updatedAt={data.updatedAt}
       additionalFiles={data.additionalFiles}
-      year={data.yearGroupTitle}
-      subject={data.subjectTitle}
-      keystage={data.keyStageTitle}
-      keystageSlug={data.keyStageSlug}
-      unit={data.unitTitle}
-      lesson={data.lessonTitle}
-      examBoardSlug={data.examBoardSlug}
-      subjectSlug={data.subjectSlug}
-      subjectParent={data.subjectParent}
-      phaseSlug={data.phaseSlug}
-      disablePupilLink={data.actions?.disablePupilShare}
-      hideSeoHelper={copyrightState.showGeoBlocked}
       useIntegratedJourneyLayout
     />
   );